refactor: extract expert review business logic into server layer and slim main routesgit add .
